+++ src/components/buttons/lan/create-config-channel.js
@@ -1,4 +1,4 @@
-const { ModalBuilder, TextInputBuilder, TextInputStyle, ActionRowBuilder, LabelBuilder } = require("discord.js");
+const { ModalBuilder, TextInputBuilder, TextInputStyle, LabelBuilder } = require("discord.js");
 
 module.exports = {
     data: {
